Extract api calls into seperate files. Move action types into seperate file.
