+++ src/App.js
@@ -1,17 +1,17 @@
-import React, { Component } from "react";
-import { Route, Redirect, Switch } from "react-router-dom";
-import { connect } from "react-redux";
-import * as actionTypes from "./store/actions/actions";
-import "materialize-css/dist/css/materialize.min.css";
-import M from "materialize-css";
+import React, { Component } from 'react';
+import { Route, Redirect, Switch } from 'react-router-dom';
+import { connect } from 'react-redux';
+import { authenticate } from './store/actions/index';
+import 'materialize-css/dist/css/materialize.min.css';
+import M from 'materialize-css';
 
-import Login from "./components/Login";
-import Register from "./components/Register";
-import Dashboard from "./components/Dashboard";
-import ProtectedRoute from "./components/ProtectedRoute";
-import Header from "./components/Header";
-import Landing from "./components/Landing";
-import BookPage from "./components/BookPage";
+import Login from './components/Login';
+import Register from './components/Register';
+import Dashboard from './components/Dashboard';
+import ProtectedRoute from './components/ProtectedRoute';
+import Header from './components/Header';
+import Landing from './components/Landing';
+import BookPage from './components/BookPage';
 
 class App extends Component {
   componentDidMount() {
@@ -48,13 +48,13 @@ class App extends Component {
 
 const mapStateToProps = state => {
   return {
-    loggedIn: state.auth.loggedIn
+    loggedIn: state.auth.loggedIn,
   };
 };
 
 const mapDispatchToProps = dispatch => {
   return {
-    authenticate: () => dispatch(actionTypes.authenticate())
+    authenticate: () => dispatch(authenticate()),
   };
 };
